In arguing that current theories of brain function cast suspicion on ESP, psychokinesis, reincarnation, and so on, I am frequently challenged with the most popular of all neuro-mythologies -- the notion that we ordinarily use only 10 percent of our brains...
This "cerebral spare tire" concept continues to nourish the clientele of
"pop psychologists" and their many recycling self-improvement schemes.  As
a metaphor for the fact that few of us fully exploit our talents, who could
deny it?  As a refuge for occultists seeking a neural basis of the miraculous,
it leaves much to be desired.
-- Barry L. Beyerstein, "The Brain and Conciousness:  Implications for
   Psi Phenomena", The Skeptical Enquirer, Vol. XII, No. 2, pg. 171